requirements.txt
aioredis==2.0.1
APScheduler==3.11.0
async-timeout==5.0.1
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
